[DEV-4096] Move matview models
