--- usaspending_api/awards/models_matviews/__init__.py
@@ -1,46 +1,0 @@
-from usaspending_api.awards.models_matviews.mv_award_summary import AwardSummaryMatview
-from usaspending_api.awards.models_matviews.mv_contract_award_search import ContractAwardSearchMatview
-from usaspending_api.awards.models_matviews.mv_directpayment_award_search import DirectPaymentAwardSearchMatview
-from usaspending_api.awards.models_matviews.mv_grant_award_search import GrantAwardSearchMatview
-from usaspending_api.awards.models_matviews.mv_idv_award_search import IDVAwardSearchMatview
-from usaspending_api.awards.models_matviews.mv_loan_award_search import LoanAwardSearchMatview
-from usaspending_api.awards.models_matviews.mv_other_award_search import OtherAwardSearchMatview
-from usaspending_api.awards.models_matviews.mv_pre2008_award_search import Pre2008AwardSearchMatview
-from usaspending_api.awards.models_matviews.subaward_view import SubawardView
-from usaspending_api.awards.models_matviews.summary_cfda_numbers_view import SummaryCfdaNumbersView
-from usaspending_api.awards.models_matviews.summary_naics_codes_view import SummaryNaicsCodesView
-from usaspending_api.awards.models_matviews.summary_psc_codes_view import SummaryPscCodesView
-from usaspending_api.awards.models_matviews.summary_state_view import SummaryStateView
-from usaspending_api.awards.models_matviews.summary_transaction_fed_acct_view import SummaryTransactionFedAcctView
-from usaspending_api.awards.models_matviews.summary_transaction_geo_view import SummaryTransactionGeoView
-from usaspending_api.awards.models_matviews.summary_transaction_month_view import SummaryTransactionMonthView
-from usaspending_api.awards.models_matviews.summary_transaction_recipient_view import SummaryTransactionRecipientView
-from usaspending_api.awards.models_matviews.summary_transaction_view import SummaryTransactionView
-from usaspending_api.awards.models_matviews.summary_view import SummaryView
-from usaspending_api.awards.models_matviews.universal_transaction_matview import UniversalTransactionView
-from usaspending_api.awards.models_matviews.vw_award_search import AwardSearchView
-
-
-__all__ = [
-    "AwardSearchView",
-    "AwardSummaryMatview",
-    "ContractAwardSearchMatview",
-    "DirectPaymentAwardSearchMatview",
-    "GrantAwardSearchMatview",
-    "IDVAwardSearchMatview",
-    "LoanAwardSearchMatview",
-    "OtherAwardSearchMatview",
-    "Pre2008AwardSearchMatview",
-    "SubawardView",
-    "SummaryCfdaNumbersView",
-    "SummaryNaicsCodesView",
-    "SummaryPscCodesView",
-    "SummaryStateView",
-    "SummaryTransactionFedAcctView",
-    "SummaryTransactionGeoView",
-    "SummaryTransactionMonthView",
-    "SummaryTransactionRecipientView",
-    "SummaryTransactionView",
-    "SummaryView",
-    "UniversalTransactionView",
-]
