+++ usaspending_api/awards/v2/filters/matview_filters.py
@@ -5,7 +5,7 @@ from django.db.models import Q
 
 from usaspending_api.accounts.views.federal_accounts_v2 import filter_on
 from usaspending_api.awards.models import FinancialAccountsByAwards
-from usaspending_api.awards.models_matviews import AwardSearchView, UniversalTransactionView
+from usaspending_api.matviews.models import AwardSearchView, UniversalTransactionView
 from usaspending_api.awards.v2.filters.filter_helpers import combine_date_range_queryset, total_obligation_queryset
 from usaspending_api.awards.v2.filters.location_filter_geocode import geocode_filter_locations
 from usaspending_api.common.exceptions import InvalidParameterException
