+++ usaspending_api/awards/v2/views/subawards.py
@@ -4,7 +4,7 @@ from django.db.models import F
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from usaspending_api.awards.models_matviews import SubawardView
+from usaspending_api.matviews.models import SubawardView
 from usaspending_api.common.cache_decorator import cache_response
 from usaspending_api.common.helpers.generic_helper import get_simple_pagination_metadata
 from usaspending_api.common.validator.pagination import PAGINATION
