+++ usaspending_api/matviews/models/mv_idv_award_search.py
@@ -1,4 +1,4 @@
-from usaspending_api.awards.models_matviews.base_award_search import BaseAwardSearchModel
+from usaspending_api.matviews.models.base_award_search import BaseAwardSearchModel
 
 
 class IDVAwardSearchMatview(BaseAwardSearchModel):
