+++ usaspending_api/search/tests/unit/test_spending_over_time_details.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from rest_framework import status
 from django_mock_queries.query import MockModel
 
-from usaspending_api.awards.models_matviews import SummaryTransactionView
+from usaspending_api.matviews.models import SummaryTransactionView
 from usaspending_api.common.helpers.unit_test_helper import add_to_mock_objects
