+++ usaspending_api/settings.py
@@ -117,6 +117,7 @@ INSTALLED_APPS = [
     "usaspending_api.download",
     "usaspending_api.bulk_download",
     "usaspending_api.recipient",
+    "usaspending_api.matviews",
     "django_spaghetti",
     "simple_history",
 ]
